chore: rename tools in tool
