+++ tool/duckduckgo/duckduckgo.go
